+++ frontend/src/api/api.ts
@@ -1,24 +1,13 @@
-// eslint-disable-next-line import/no-cycle
 import { setError } from '../state/reducers/apiErrorReducer';
+import store from '../state/store';
+import { ApiError } from '../types';
 
 const API_BASE = 'http://localhost:3001/api';
-
-export class ApiError extends Error {
-  status: string;
-  body: any;
-
-  constructor(res: any, body: any) {
-    super(body.message || res.statusText);
-    this.status = res.status;
-    this.name = 'ApiError';
-    this.body = body;
-  }
-}
 
 const responseHandler = async (res: Response): Promise<any> => {
   if (!res.ok) {
     const body = await res.json();
-    setError(new ApiError(res, body));
+    store.dispatch(setError(new ApiError(res, body)));
   }
 
   const json = await res.json();
