+++ frontend/src/state/reducers/apiErrorReducer.ts
@@ -1,23 +1,17 @@
-/* eslint-disable import/no-cycle */
-import { ApiError } from '../../api/api';
-import { ErrorState, ErrorStateAction } from '../../types';
-import store from '../store';
+import { ApiError, CLEAR_ERROR, ErrorState, ErrorStateAction, SET_ERROR } from '../../types';
 
 const defaultState: ErrorState = {
   apiError: null,
 };
 
-export const SET_ERROR = 'SET_ERROR';
-export const CLEAR_ERROR = 'CLEAR_ERROR';
-
 const reducer = (state: ErrorState = defaultState, action: ErrorStateAction): ErrorState => {
   switch (action.type) {
-    case 'SET_ERROR':
+    case SET_ERROR:
       return {
         ...state,
         apiError: action.payload,
       };
-    case 'CLEAR_ERROR':
+    case CLEAR_ERROR:
       return {
         ...state,
         apiError: null,
@@ -27,15 +21,11 @@ const reducer = (state: ErrorState = defaultState, action: ErrorStateAction): Er
   }
 };
 
-export const setError = (error: ApiError): void => {
-  store.dispatch({
-    type: 'SET_ERROR',
-    payload: error,
-  });
-};
+export const setError = (error: ApiError): ErrorStateAction => ({
+  type: 'SET_ERROR',
+  payload: error,
+});
 
-export const clearError = (): void => {
-  store.dispatch({ type: 'CLEAR_ERROR' });
-};
+export const clearError = (): ErrorStateAction => ({ type: 'CLEAR_ERROR' });
 
 export default reducer;
