+++ frontend/src/state/store.ts
@@ -1,5 +1,4 @@
 import { createStore } from 'redux';
-// eslint-disable-next-line import/no-cycle
 import reducer from './reducers/apiErrorReducer';
 
 const store = createStore(reducer);
